Lab 3/requirements.txt
psycopg2-binary==2.9.5
requests==2.28.2
py7zr==0.20.4
flask==2.2.4
flask_sqlalchemy==3.0.3
flask_migrate==4.0.4

